move list view back to the index
